# activity/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Activity Service API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/sessions/{sessionId}/attendees": {
            "get": {
                "security": [
                    {
                        "BearerAuth": [],
                        "GatewayKeyAuth": []
                    }
                ],
                "description": "Lists all attendees/RSVPs for a specific session with optional filters",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "RSVPs"
                ],
                "summary": "List attendees for a session",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Session ID",
                        "name": "sessionId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Filter by status (going, pending, not_going, maybe, promoted)",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Limit number of results (default 20)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset for pagination (default 0)",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Attendees retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/api.Response-github_com_rasparac_rekreativko-api_activity_internal_interfaces_http_dtos_AttendeeListResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    }
                }
            }
        },
        "/api/v1/sessions/{sessionId}/rsvp": {
            "get": {
                "security": [
                    {
                        "BearerAuth": [],
                        "GatewayKeyAuth": []
                    }
                ],
                "description": "Retrieves the authenticated user's RSVP for a specific session",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "RSVPs"
                ],
                "summary": "Get user's RSVP",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Session ID",
                        "name": "sessionId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "RSVP retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/api.Response-github_com_rasparac_rekreativko-api_activity_internal_interfaces_http_dtos_AttendeeResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    },
                    "404": {
                        "description": "RSVP not found",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": [],
                        "GatewayKeyAuth": []
                    }
                ],
                "description": "Updates the authenticated user's RSVP for a specific session",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "RSVPs"
                ],
                "summary": "Update an RSVP",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Session ID",
                        "name": "sessionId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated RSVP data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_rasparac_rekreativko-api_activity_internal_interfaces_http_dtos.UpdateRSVPRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "RSVP updated successfully",
                        "schema": {
                            "$ref": "#/definitions/api.Response-github_com_rasparac_rekreativko-api_activity_internal_interfaces_http_dtos_AttendeeResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    },
                    "404": {
                        "description": "RSVP not found",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": [],
                        "GatewayKeyAuth": []
                    }
                ],
                "description": "Creates a new RSVP for the authenticated user for a specific session",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "RSVPs"
                ],
                "summary": "Create an RSVP for a session",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Session ID",
                        "name": "sessionId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "RSVP data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_rasparac_rekreativko-api_activity_internal_interfaces_http_dtos.CreateRSVPRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "RSVP created successfully",
                        "schema": {
                            "$ref": "#/definitions/api.Response-github_com_rasparac_rekreativko-api_activity_internal_interfaces_http_dtos_CreateRSVPResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    },
                    "409": {
                        "description": "Already RSVPed or session full",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": [],
                        "GatewayKeyAuth": []
                    }
                ],
                "description": "Cancels the authenticated user's RSVP for a specific session",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "RSVPs"
                ],
                "summary": "Cancel an RSVP",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Session ID",
                        "name": "sessionId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "RSVP cancelled successfully",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    },
                    "404": {
                        "description": "RSVP not found",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api.APIError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "invalid_request"
                },
                "details": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "message": {
                    "type": "string",
                    "example": "Invalid request"
                }
            }
        },
        "api.Response-any": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "only on success"
                },
                "error": {
                    "description": "only on error",
                    "allOf": [
                        {
                            "$ref": "#/definitions/api.APIError"
                        }
                    ]
                },
                "message": {
                    "type": "string",
                    "example": "Operation completed successfully"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                },
                "timestamp": {
                    "type": "string",
                    "example": "2020-01-01T00:00:00Z"
                }
            }
        },
...
        "api.Response-github_com_rasparac_rekreativko-api_activity_internal_interfaces_http_dtos_AttendeeListResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "only on success",
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_rasparac_rekreativko-api_activity_internal_interfaces_http_dtos.AttendeeListResponse"
                        }
                    ]
                },
                "error": {
                    "description": "only on error",
                    "allOf": [
                        {
                            "$ref": "#/definitions/api.APIError"
                        }
                    ]
                },
                "message": {
                    "type": "string",
                    "example": "Operation completed successfully"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                },
                "timestamp": {
                    "type": "string",
                    "example": "2020-01-01T00:00:00Z"
                }
            }
        },
        "api.Response-github_com_rasparac_rekreativko-api_activity_internal_interfaces_http_dtos_AttendeeResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "only on success",
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_rasparac_rekreativko-api_activity_internal_interfaces_http_dtos.AttendeeResponse"
                        }
                    ]
                },
                "error": {
                    "description": "only on error",
                    "allOf": [
                        {
                            "$ref": "#/definitions/api.APIError"
                        }
                    ]
                },
                "message": {
                    "type": "string",
                    "example": "Operation completed successfully"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                },
                "timestamp": {
                    "type": "string",
                    "example": "2020-01-01T00:00:00Z"
                }
            }
        },
...
        "api.Response-github_com_rasparac_rekreativko-api_activity_internal_interfaces_http_dtos_CreateRSVPResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "only on success",
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_rasparac_rekreativko-api_activity_internal_interfaces_http_dtos.CreateRSVPResponse"
                        }
                    ]
                },
                "error": {
                    "description": "only on error",
                    "allOf": [
                        {
                            "$ref": "#/definitions/api.APIError"
                        }
                    ]
                },
                "message": {
                    "type": "string",
                    "example": "Operation completed successfully"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                },
                "timestamp": {
                    "type": "string",
                    "example": "2020-01-01T00:00:00Z"
                }
            }
        },
...
        "github_com_rasparac_rekreativko-api_activity_internal_interfaces_http_dtos.AttendeeListResponse": {
            "type": "object",
            "properties": {
                "attendees": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_rasparac_rekreativko-api_activity_internal_interfaces_http_dtos.AttendeeResponse"
                    }
                },
                "limit": {
                    "type": "integer"
                },
                "offset": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "github_com_rasparac_rekreativko-api_activity_internal_interfaces_http_dtos.AttendeeResponse": {
            "type": "object",
            "properties": {
                "activity_group_id": {
                    "type": "string",
                    "example": "123e4567-e89b-12d3-a456-426655440000"
                },
                "created_at": {
                    "type": "string",
                    "example": "2024-01-01T00:00:00Z"
                },
                "id": {
                    "type": "string",
                    "example": "123e4567-e89b-12d3-a456-426655440000"
                },
                "session_id": {
                    "type": "string",
                    "example": "123e4567-e89b-12d3-a456-426655440000"
                },
                "source": {
                    "type": "string",
                    "example": "rsvp_manual"
                },
                "status": {
                    "type": "string",
                    "example": "going"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2024-01-01T00:00:00Z"
                },
                "user_id": {
                    "type": "string",
                    "example": "123e4567-e89b-12d3-a456-426655440000"
                }
            }
        },
...
        "github_com_rasparac_rekreativko-api_activity_internal_interfaces_http_dtos.CreateRSVPRequest": {
            "type": "object",
            "required": [
                "status"
            ],
            "properties": {
                "status": {
                    "type": "string",
                    "enum": [
                        "going",
                        "not_going",
                        "maybe"
                    ],
                    "example": "going"
                }
            }
        },
        "github_com_rasparac_rekreativko-api_activity_internal_interfaces_http_dtos.CreateRSVPResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string",
                    "example": "123e4567-e89b-12d3-a456-426655440000"
                }
            }
        },
...
        "github_com_rasparac_rekreativko-api_activity_internal_interfaces_http_dtos.UpdateRSVPRequest": {
            "type": "object",
            "required": [
                "status"
            ],
            "properties": {
                "status": {
                    "type": "string",
                    "enum": [
                        "going",
                        "not_going",
                        "maybe"
                    ],
                    "example": "not_going"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "GatewayKeyAuth": {
            "description": "Key added automatically by the gateway when proxying requests; required directly here only when bypassing the gateway.",
            "type": "apiKey",
            "name": "X-Gateway-Key",
            "in": "header"
        }
    },
    "security": [
        {
            "GatewayKeyAuth": []
        }
    ]
}